Move quick_date_picker.dart to lib/presentation/widgets/search
